Revert "changes for permission"
This reverts commit 3b50453f2e2d96cc9b7ad4316a5f7f1c23f8b690.

+++ src/ApplicationCore/Models/PermissionModel.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int RoleId { get; set; }
-        public int ModuleId { get; set; }
+        public decimal ModuleId { get; set; }
         public string ModuleName { get; set; }
         public string Url { get; set; }
         public string Icon { get; set; }
